Change purification step
When purification was triggered due to C-level DIF, it could not re-run the MHfun function because columns were identified by column name. When the C-level items were removed, some column names would go missing, causing the function to terminate with an error. This was fixed by changing the column identification from its name to the column number so that identification is no longer dependent on the name but only on the position.